feat(docs): /docs URI for swagger-ui & refactor front packages
