+++ application/src/main/java/fi/hsl/parkandride/front/BBox.java
@@ -34,4 +34,35 @@ public class BBox {
                 max.getLon(), min.getLat(),
                 min.getLon(), min.getLat()));
     }
+
+    public static class Point {
+
+        private double lon;
+
+        private double lat;
+
+        public Point() {}
+
+        public Point(double lon, double lat) {
+            this.lon = lon;
+            this.lat = lat;
+        }
+
+        public double getLon() {
+            return lon;
+        }
+
+        public void setLon(double lon) {
+            this.lon = lon;
+        }
+
+        public double getLat() {
+            return lat;
+        }
+
+        public void setLat(double lat) {
+            this.lat = lat;
+        }
+
+    }
 }
