--- application/src/main/java/fi/hsl/parkandride/front/Point.java
@@ -1,32 +1,0 @@
-package fi.hsl.parkandride.front;
-
-public class Point {
-
-    private double lon;
-
-    private double lat;
-
-    public Point() {}
-
-    public Point(double lon, double lat) {
-        this.lon = lon;
-        this.lat = lat;
-    }
-
-    public double getLon() {
-        return lon;
-    }
-
-    public void setLon(double lon) {
-        this.lon = lon;
-    }
-
-    public double getLat() {
-        return lat;
-    }
-
-    public void setLat(double lat) {
-        this.lat = lat;
-    }
-
-}
